Introduction

Compute Legendre polynomials, Associated Legendre polynomials, and their derivatives using a 3-term recursion relation (Bonnet’s recursion formula).

\[P_\ell(x) = \left((2\ell-1) x P_{\ell-1}(x) - (\ell-1)P_{\ell - 2}(x)\right)/\ell\]

Currently this package evaluates the standard polynomials that satisfy $P_\ell(1) = 1$ and $P_0(x) = 1$. These are normalized as

\[\int_{-1}^1 P_m(x) P_n(x) dx = \frac{2}{2n+1} \delta_{mn}.\]

There are six main functions:

Pl(x,l): this evaluates the Legendre polynomial for a given degree l at the argument x. The argument needs to satisfy -1 <= x <= 1.

collectPl(x; lmax): this evaluates all the polynomials for l lying in 0:lmax at the argument x. As before the argument needs to lie in the domain of validity. Functionally this is equivalent to Pl.(x, 0:lmax), except collectPl evaluates the result in one pass, and is therefore faster. There is also the in-place version collectPl! that uses a pre-allocated array.

Plm(x, l, m): this evaluates the associated Legendre polynomial $P_\ell,m(x)$ at the argument $x$. The argument needs to satisfy -1 <= x <= 1.

collectPlm(x; m, lmax): this evaluates the associated Legendre polynomials with coefficient m for l = 0:lmax. There is also an in-place version collectPlm! that uses a pre-allocated array.

dnPl(x, l, n): this evaluates the $n$-th derivative of the Legendre polynomial $P_\ell(x)$ at the argument $x$. The argument needs to satisfy -1 <= x <= 1.

collectdnPl(x; n, lmax): this evaluates the $n$-th derivative of all the Legendre polynomials for l = 0:lmax. There is also an in-place version collectdnPl! that uses a pre-allocated array.

Quick Start

Evaluate the Legendre polynomial for one l at an argumentx as Pl(x, l):

julia> Pl(0.5, 3)
-0.4375

Evaluate the associated Legendre Polynomial one l,m pair as Plm(x, l, m):

julia> Plm(0.5, 3, 2)
5.625

Evaluate the nth derivative for one l as dnPl(x, l, n):

julia> dnPl(0.5, 3, 2)
7.5

Evaluate all the polynomials for l in 0:lmax as collectPl(x; lmax)

julia> collectPl(0.5, lmax = 3)
4-element OffsetArray(::Vector{Float64}, 0:3) with eltype Float64 with indices 0:3:
  1.0
  0.5
 -0.125
 -0.4375

Evaluate all the associated Legendre Polynomials for coefficient m as collectPlm(x; lmax, m):

julia> collectPlm(0.5, lmax = 5, m = 3)
6-element OffsetArray(::Vector{Float64}, 0:5) with eltype Float64 with indices 0:5:
   0.0
   0.0
   0.0
  -9.742785792574933
 -34.099750274012266
 -42.62468784251533

Evaluate all the nth derivatives as collectdnPl(x; lmax, n):

julia> collectdnPl(0.5, lmax = 5, n = 3)
6-element OffsetArray(::Vector{Float64}, 0:5) with eltype Float64 with indices 0:5:
  0.0
  0.0
  0.0
 15.0
 52.5
 65.625

Increase precision

The precision of the result may be changed by using arbitrary-precision types such as BigFloat. For example, using Float64 arguments we obtain

julia> Pl(1/3, 5)
0.33333333333333337

whereas using BigFloat, we obtain

julia> Pl(big(1)/3, 5)
0.3333333333333333333333333333333333333333333333333333333333333333333333333333305

The precision of the latter may be altered using setprecision, as

julia> setprecision(300) do
       Pl(big(1)/3, 5)
       end
0.33333333333333333333333333333333333333333333333333333333333333333333333333333333333333333317

This is particularly important to avoid overflow while computing high-order derivatives. For example:

julia> dnPl(0.5, 300, 200) # Float64
NaN

julia> dnPl(big(1)/2, 300, 200) # BigFloat
1.738632750542319394663553898425873258768856732308227932150592526951212145232716e+499

Reference

LegendrePolynomials.LegendrePolynomialIterator — Type

LegendrePolynomialIterator(x, [lmax::Integer])

Return an iterator that generates the values of the Legendre polynomials $P_\ell(x)$ for the given x. If lmax is specified then only the values of $P_\ell(x)$ from 0 to lmax are returned.

Warn

LegendrePolynomialIterator will not be a part of the public API from the next minor release.

Examples

julia> import LegendrePolynomials: LegendrePolynomialIterator

julia> iter = LegendrePolynomialIterator(0.5, 4);

julia> collect(iter)
5-element OffsetArray(::Vector{Float64}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625

julia> iter = LegendrePolynomialIterator(0.5);

julia> collect(Iterators.take(iter, 5)) # evaluete 5 elements (l = 0:4)
5-element Vector{Float64}:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625

julia> collect(Iterators.take(Iterators.drop(iter, 100), 5)) # evaluate Pl for l = 100:104
5-element Vector{Float64}:
 -0.0605180259618612
  0.02196749072249231
  0.08178451892628381
  0.05963329258495025
 -0.021651535258316177

LegendrePolynomials.Pl — Method

Pl(x, l::Integer)

Compute the Legendre Polynomial $P_\ell(x)$ for the argument x and the degree l.

Examples

julia> Pl(1, 2)
1.0

julia> Pl(0.5, 20)
-0.04835838106737356

LegendrePolynomials.Plm — Function

Plm(x, l::Integer, m::Integer, [cache::AbstractVector])

Compute the associatedLegendre polynomial $P_\ell^m(x)$. Optionally a pre-allocated vector cache may be provided, which must have a minimum length of l - m + 1 and may be overwritten during the computation.

The polynomials are defined to include the Condon-Shortley phase $(-1)^m$.

The coefficient m must be non-negative. For m == 0 this function just returns Legendre polynomials.

Examples

julia> Plm(0.5, 3, 2)
5.625

julia> Plm(0.5, 4, 0) == Pl(0.5, 4)
true

LegendrePolynomials.collectPl! — Method

collectPl!(v::AbstractVector, x; [lmax::Integer = length(v) - 1])

Compute the Legendre Polynomials $P_\ell(x)$ for the argument x and all degrees l in 0:lmax, and store them in v.

At output v[firstindex(v) + l] == Pl(x,l).

Examples

julia> v = zeros(4);

julia> collectPl!(v, 0.5)
4-element Vector{Float64}:
  1.0
  0.5
 -0.125
 -0.4375

julia> v = zeros(0:4);

julia> collectPl!(v, 0.5, lmax = 3) # only l from 0 to 3 are populated
5-element OffsetArray(::Vector{Float64}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
  0.0

LegendrePolynomials.collectPl — Method

collectPl(x; lmax::Integer)

Compute the Legendre Polynomial $P_\ell(x)$ for the argument x and all degrees l in 0:lmax. Return P with indices 0:lmax, where P[l] == Pl(x,l)

Examples

julia> collectPl(0.5, lmax = 4)
5-element OffsetArray(::Vector{Float64}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625

LegendrePolynomials.collectPlm! — Method

collectPlm!(v::AbstractVector, x; lmax::Integer, m::Integer)

Compute the associated Legendre Polynomial $P_\ell^m(x)$ for the argument x and all degrees l = 0:lmax, and store the result in v. The polynomials for l < m are defined to be zero.

The polynomials are defined to include the Condon-Shortley phase $(-1)^m$.

The coefficient m must be greater than or equal to zero.

At output, v[l + firstindex(v)] == Plm(x, l, m) for l = 0:lmax.

Examples

julia> v = zeros(4);

julia> collectPlm!(v, 0.5, lmax = 3, m = 2)
4-element Vector{Float64}:
 0.0
 0.0
 2.25
 5.625

LegendrePolynomials.collectPlm — Method

collectPlm(x; lmax::Integer, m::Integer)

Compute the associated Legendre Polynomial $P_\ell^m(x)$ for the argument x and all degrees l = 0:lmax. The polynomials for l < m are defined to be zero.

The polynomials are defined to include the Condon-Shortley phase $(-1)^m$.

The coefficient m must be greater than or equal to zero.

Returns v with indices 0:lmax, where v[l] == Plm(x, l, m).

Examples

julia> collectPlm(0.5, lmax = 3, m = 2)
4-element OffsetArray(::Vector{Float64}, 0:3) with eltype Float64 with indices 0:3:
 0.0
 0.0
 2.25
 5.625

LegendrePolynomials.collectdnPl! — Method

collectdnPl!(v::AbstractVector, x; lmax::Integer, n::Integer)

Compute the $n$-th derivative of a Legendre Polynomial $P_\ell(x)$ for the argument x and all degrees l = 0:lmax, and store the result in v.

The order of the derivative n must be greater than or equal to zero.

At output, v[l + firstindex(v)] == dnPl(x, l, n) for l = 0:lmax.

Examples

julia> v = zeros(4);

julia> collectdnPl!(v, 0.5, lmax = 3, n = 2)
4-element Vector{Float64}:
 0.0
 0.0
 3.0
 7.5

LegendrePolynomials.collectdnPl — Method

collectdnPl(x; lmax::Integer, n::Integer)

Compute the $n$-th derivative of a Legendre Polynomial $P_\ell(x)$ for the argument x and all degrees l = 0:lmax.

The order of the derivative n must be greater than or equal to zero.

Returns v with indices 0:lmax, where v[l] == dnPl(x, l, n).

Examples

julia> collectdnPl(0.5, lmax = 3, n = 2)
4-element OffsetArray(::Vector{Float64}, 0:3) with eltype Float64 with indices 0:3:
 0.0
 0.0
 3.0
 7.5

LegendrePolynomials.dnPl — Function

dnPl(x, l::Integer, n::Integer, [cache::AbstractVector])

Compute the $n$-th derivative $d^n P_\ell(x)/dx^n$ of the Legendre polynomial $P_\ell(x)$. Optionally a pre-allocated vector cache may be provided, which must have a minimum length of l - n + 1 and may be overwritten during the computation.

The order of the derivative n must be non-negative. For n == 0 this function just returns Legendre polynomials.

Examples

julia> dnPl(0.5, 3, 2) # second derivative of P3(x) at x = 0.5
7.5

julia> dnPl(0.5, 4, 0) == Pl(0.5, 4) # zero-th order derivative == Pl(x)
true