LegendrePolynomials.jl

LegendrePolynomials.Pl! — Method

Pl!(arr::AbstractArray, x::Real; [lmax = maximum(axes(arr,1))])

Computes the Legendre Polynomials $P_l(x)$ for the argument x and l = 0:lmax, and saves it in arr. Assumes that arr has 0-based indexing.

All dimensions of arr should have indices starting from 0. The axes of arr should be (0:l, 0:n) where l >= lmax and n >= 0

At output, arr[l,0] == Pl(x,l)

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.Pl — Method

Pl(x::Real, l::Integer)

Computes the Legendre Polynomials $P_l(x)$ for the argument x and the degree l

LegendrePolynomials.Pl — Method

Pl(x::Real; lmax::Integer)

Computes the Legendre Polynomials $P_l(x)$ for the argument x and l = 0:lmax. Returns an OffsetArray P with indices 0:lmax, where P[l] == Pl(x,l)

LegendrePolynomials.Pl_d2Pl! — Method

Pl_d2Pl!(arr::AbstractMatrix, x::Real; [lmax = maximum(axes(arr,1))])

Computes the first and second derivatives $d_xP_l(x)$ and $d_x^2P_l(x)$, for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

The first dimension of arr should be 0:l, where l >= lmax The second dimension should be 0:n where n >= 2

At output, arr[l,0] == dPl(x,l) and arr[l,2] == d2Pl(x,l)

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.Pl_d2Pl — Method

Pl_d2Pl(x::Real, l::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their second derivatives $d_x^2P_l(x)$ for the argument x and the degree l

LegendrePolynomials.Pl_d2Pl — Method

Pl_d2Pl(x::Real; lmax::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their second derivatives $d_x^2P_l(x)$ for the argument x and l = 0:lmax

Returns OffsetArrays P and d2P with indices 0:lmax, where P[l] == Pl(x,l) and d2P[l] == d2Pl(x,l)

LegendrePolynomials.Pl_dPl! — Method

Pl_dPl!(arr, x; [lmax = maximum(axes(arr,1))])

Computes the Legendre Polynomials $P_l(x)$ and their derivatives $d_xP_l(x)$ for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

All dimensions of arr should have indices starting from 0 The axes of arr should be (0:l, 0:n) where l >= lmax and $n >= 1$

At output, arr[l,0] == Pl(x,l) and arr[l,1] == dPl(x,l)

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.Pl_dPl — Method

Pl_dPl(x::Real, l::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their first derivatives $d_xP_l(x)$ for the argument x and the degree l.

LegendrePolynomials.Pl_dPl — Method

Pl_dPl(x::Real; lmax::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their derivatives $d_xP_l(x)$ for the argument x and l = 0:lmax Returns OffsetArrays P and dP with indices 0:lmax, where P[l] == Pl(x,l) and dP[l] == dPl(x,l).

LegendrePolynomials.Pl_dPl_d2Pl! — Method

Pl_dPl_d2Pl!(arr::AbstractMatrix, x::Real; [lmax = maximum(axes(arr,1))])

Computes the Legendre Polynomials $P_l(x)$ and their first and second derivatives $d_xP_l(x)$ and $d_x^2P_l(x)$, for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

All dimensions of arr should have indices starting from 0 The axes of arr should be (0:l, 0:n) where l >= lmax and n >= 2

At output, arr[l,0] == Pl(x,l), arr[l,1] == dPl(x,l) and arr[l,2] == d2Pl(x,l)

The optional keyword argument lmax can specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.Pl_dPl_d2Pl — Method

Pl_dPl_d2Pl(x::Real, l::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their first derivatives $d_xP_l(x)$ and second derivatives $d_x^2P_l(x)$ for the argument $x$ and the degree $l$

LegendrePolynomials.Pl_dPl_d2Pl — Method

Pl_dPl_d2Pl(x::Real; lmax::Integer)

Computes the the Legendre Polynomials $P_l(x)$ and their first derivatives $d_xP_l(x)$ and second derivatives $d_x^2P_l(x)$ for the argument x and l = 0:lmax. Returns OffsetArrays P, dP and d2P with indices 0:lmax, where P[l] == Pl(x,l), dP[l] == dPl(x,l) and d2P[l] == d2Pl(x,l)

LegendrePolynomials.d2Pl! — Method

d2Pl!(arr::AbstractArray, x::Real; [lmax = maximum(axes(arr,1))])

Computes the second derivatives of Legendre Polynomials $d_x^2P_l(x)$ for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

The first dimension of arr should be 0:l, where l >= lmax The second dimension – if arr is 2D – should contain the index 2

At output, arr[l] == d2Pl(x,l) if arr is a Vector or arr[l,2] == d2Pl(x,l) if arr is a Matrix

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.d2Pl — Method

d2Pl(x::Real, l::Integer)

Computes the second derivatives of Legendre Polynomials $d_x^2P_l(x)$ for the argument x and the degree l

LegendrePolynomials.d2Pl — Method

d2Pl(x::Real; lmax::Integer)

Computes the second derivatives of Legendre Polynomials $d_x^2P_l(x)$ for the argument x and l = 0:lmax Returns an OffsetArray d2P with indices 0:lmax, where d2P[l] == d2Pl(x,l)

LegendrePolynomials.dPl! — Method

dPl!(arr::AbstractArray, x::Real; [lmax = maximum(axes(arr,1))])

Computes the first derivatives of Legendre Polynomials $d_xP_l(x)$ for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

The first dimension of arr should be 0:l, where l >= lmax The second dimension – if arr is 2D – should contain the index 1

At output, arr[l] == dPl(x,l) if arr is a Vector or arr[l,1] == dPl(x,l) if arr is a Matrix

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.dPl — Method

dPl(x::Real, l::Integer)

Computes the first derivatives of Legendre Polynomials $d_xP_l(x)$ for the argument x and the degree l

LegendrePolynomials.dPl — Method

dPl(x::Real; lmax::Integer)

Computes the first derivatives of Legendre Polynomials $d_xP_l(x)$ for the argument x and l = 0:lmax. Returns an OffsetArray dP with indices 0:lmax, where dP[l] = dPl(x,l)

LegendrePolynomials.dPl_d2Pl! — Method

dPl_d2Pl!(arr::AbstractMatrix, x::Real; [lmax = maximum(axes(arr,1))])

Computes the first and second derivatives $d_xP_l(x)$ and $d_x^2P_l(x)$, for the argument x and l = 0:lmax, and saves them in arr. Assumes that arr has 0-based indexing.

The first dimension of arr should be 0:l, where l >= lmax The second dimension should be 1:n where n >= 2

At output, arr[l,1] == dPl(x,l) and arr[l,2] == d2Pl(x,l)

The optional keyword argument lmax may specify the range of $l$'s to compute. It defaults to lmax = maximum(axes(arr,1))

LegendrePolynomials.dPl_d2Pl — Method

dPl_d2Pl(x::Real, l::Integer)

Computes the the first and second derivatives of Legendre Polynomials $d_xP_l(x)$ and $d_x^2P_l(x)$ for the argument x and the degree l

LegendrePolynomials.dPl_d2Pl — Method

dPl_d2Pl(x::Real; lmax::Integer)

Computes the the first and second derivatives of Legendre Polynomials $d_xP_l(x)$ and $d_x^2P_l(x)$ for the argument x and l = 0:lmax

Returns OffsetArrays dP and d2P with indices 0:lmax, where dP[l] == dPl(x,l) and d2P[l] == d2Pl(x,l)