Introduction

Compute Legendre polynomials using a 3-term recursion relation (Bonnet’s recursion formula).

\[P_\ell(x) = \left((2\ell-1) x P_{\ell-1}(x) - (\ell-1)P_{\ell - 2}(x)\right)/\ell\]

Currently this package evaluates the standard polynomials that satisfy $P_\ell(1) = 1$. These are normalized as

\[\int P_m(x) P_n(x) dx = \frac{2}{2n+1} \delta_{mn}.\]

There are two main functions:

Pl(x,l): this evaluates the Legendre polynomial for a given degree l at the argument x. The argument needs to satisfy -1 <= x <= 1.

collectPl(x; lmax): this evaluates all the polynomials for l lying in 0:lmax at the argument x. As before the argument needs to lie in the domain of validity. Functionally this is equivalent to Pl.(x, 0:lmax), except collectPl evaluates the result in one pass, and is therefore faster. There is also the in-place version collectPl! that uses a pre-allocated array.

Quick Start

Evaluate the Legendre polynomial for one l at an argumentx as Pl(x, l):

julia> Pl(0.5, 3)
-0.4375

Evaluate all the polynomials for l in 0:lmax as

julia> collectPl(0.5, lmax = 3)
4-element OffsetArray(::Array{Float64,1}, 0:3) with eltype Float64 with indices 0:3:
  1.0
  0.5
 -0.125
 -0.4375

Increase precision

The precision of the result may be changed by using arbitrary-precision types such as BigFloat. For example, using Float64 arguments we obtain

julia> Pl(1/3, 5)
0.33333333333333337

whereas using BigFloat, we obtain

julia> Pl(big(1)/3, 5)
0.3333333333333333333333333333333333333333333333333333333333333333333333333333305

The precision of the latter may be altered using setprecision, as

julia> setprecision(300) do 
       Pl(big(1)/3, 5)
       end
0.33333333333333333333333333333333333333333333333333333333333333333333333333333333333333333317

Reference

LegendrePolynomials.LegendrePolynomialIterator — Type

LegendrePolynomialIterator(x, [lmax::Integer])

Return an iterator that generates the values of the Legendre polynomials $P_l(x)$ for the given x. If lmax is specified then only the values of $P_l(x)$ from 0 to lmax are returned.

Examples

julia> import LegendrePolynomials: LegendrePolynomialIterator

julia> iter = LegendrePolynomialIterator(0.5, 4);

julia> collect(iter)
5-element OffsetArray(::Array{Float64,1}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625

julia> iter = LegendrePolynomialIterator(0.5);

julia> collect(Iterators.take(iter, 5)) # evaluete 5 elements (l = 0:4)
5-element Array{Float64,1}:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625

julia> collect(Iterators.take(Iterators.drop(iter, 100), 5)) # evaluate Pl for l = 100:104
5-element Array{Float64,1}:
 -0.0605180259618612
  0.02196749072249231
  0.08178451892628381
  0.05963329258495025
 -0.021651535258316177

LegendrePolynomials.Pl — Method

Pl(x, l::Integer)

Compute the Legendre Polynomial $P_l(x)$ for the argument x and the degree l

Examples

julia> Pl(1, 2)
1.0

julia> Pl(0.5, 20)
-0.04835838106737356

LegendrePolynomials.collectPl! — Method

collectPl!(v::AbstractVector, x; [lmax::Integer = length(v) - 1])

Compute the Legendre Polynomials $P_l(x)$ for the argument x and all degrees l in 0:lmax, and store them in v.

At output v[firstindex(v) + l] == Pl(x,l).

Examples

julia> v = zeros(4);

julia> collectPl!(v, 0.5)
4-element Array{Float64,1}:
  1.0
  0.5
 -0.125
 -0.4375

julia> v = zeros(0:4);

julia> collectPl!(v, 0.5, lmax = 3) # only l from 0 to 3 are populated
5-element OffsetArray(::Array{Float64,1}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
  0.0

LegendrePolynomials.collectPl — Method

collectPl(x; lmax::Integer)

Compute the Legendre Polynomial $P_l(x)$ for the argument x and all degrees l in 0:lmax. Return an OffsetArray P with indices 0:lmax, where P[l] == Pl(x,l)

Examples

julia> collectPl(0.5, lmax = 4)
5-element OffsetArray(::Array{Float64,1}, 0:4) with eltype Float64 with indices 0:4:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625